Rewrite disk_health in python and change how offline repo works

+++ create_overlays.sh
@@ -5,21 +5,37 @@
 
 set -eu
 
+NO_UNPACK=false
+
+while getopts ":n" opt; do
+    case "$opt" in
+        n)
+            NO_UNPACK=true
+        ;;
+        \?)
+            echo "Unkown option"
+        ;;
+    esac
+done
+shift $((OPTIND - 1))
+
 ./build_client_packages.sh
 
 cd client
 
-rm -rf overlays/x86_64
-rm -rf overlays/x86
-
-mkdir -p overlays/x86_64/home/ssh
-mkdir -p overlays/x86/home/ssh
-
-mkdir -p ../server/package/boot/x86_64
-mkdir -p ../server/package/boot/x86
-
-tar -xzf overlays/x86_64.apkovl.tar.gz -C overlays/x86_64
-tar -xzf overlays/x86.apkovl.tar.gz -C overlays/x86
+if [ "$NO_UNPACK" = false ]; then
+    rm -rf overlays/x86_64
+    rm -rf overlays/x86
+    
+    mkdir -p overlays/x86_64/home/ssh
+    mkdir -p overlays/x86/home/ssh
+    
+    mkdir -p ../server/package/boot/x86_64
+    mkdir -p ../server/package/boot/x86
+    
+    tar -xzf overlays/x86_64.apkovl.tar.gz -C overlays/x86_64
+    tar -xzf overlays/x86.apkovl.tar.gz -C overlays/x86
+fi
 
 # Remove the existing home directory. (everything within will be replaced)
 rm -rf overlays/x86_64/home/ssh/*
